+++ coloraide/color.py
@@ -710,7 +710,11 @@ class Color(metaclass=ColorMeta):
             self._space._serialize()[0],
             ' '.join(
                 [
-                    util.fmt_float(self[i], util.DEF_PREC, util.DEF_ROUND_MODE) + POSTFIX[channels[i].angle]
+                    util.fmt_float(
+                        self[i],
+                        util.DEF_PREC,
+                        util.DEF_ROUND_MODE
+                    ) + POSTFIX[channels[i].angle if channels[i].angle and not math.isnan(self[i]) else ANGLE_NULL]
                     for i in range(l - 1)
                 ]
             ),