+++ coloraide/interpolate/continuous.py
@@ -8,7 +8,7 @@ from typing import Any
 
 
 def adjust_shorter(h1: float, h2: float, offset: float, mx: float) -> tuple[float, float]:
-    """Adjust the given hues."""
+    """Adjust the given hues such that they favor the shorter arc."""
 
     half = mx / 2
     d = h2 - h1
@@ -22,7 +22,7 @@ def adjust_shorter(h1: float, h2: float, offset: float, mx: float) -> tuple[floa
 
 
 def adjust_longer(h1: float, h2: float, offset: float, mx: float) -> tuple[float, float]:
-    """Adjust the given hues."""
+    """Adjust the given hues such that they favor the longer arc."""
 
     half = mx / 2
     d = h2 - h1
@@ -36,7 +36,7 @@ def adjust_longer(h1: float, h2: float, offset: float, mx: float) -> tuple[float
 
 
 def adjust_increase(h1: float, h2: float, offset: float, mx: float) -> tuple[float, float]:
-    """Adjust the given hues."""
+    """Adjust the given hues such that they are increasing."""
 
     if h2 < h1:
         h2 += mx
@@ -45,7 +45,7 @@ def adjust_increase(h1: float, h2: float, offset: float, mx: float) -> tuple[flo
 
 
 def adjust_decrease(h1: float, h2: float, offset: float, mx: float) -> tuple[float, float]:
-    """Adjust the given hues."""
+    """Adjust the given hues such that they are decreasing."""
 
     if h2 > h1:
         h2 -= mx
@@ -123,6 +123,8 @@ class InterpolatorContinuous(Interpolator[AnyColor]):
         coords = self.coordinates
         end = self.length - 2
         hue_index = self.hue_index
+        is_longer = self.hue == 'longer'
+        mx = self.max_hue
 
         # Normalize hue
         offset = 0.0
@@ -183,8 +185,16 @@ class InterpolatorContinuous(Interpolator[AnyColor]):
                     # else just backfill the current known value.
                     point = 1 / (x - backfill + 1)
                     for e, c in enumerate(coords[backfill:x], 1):
-                        p = alg.lerp(last, b, point * e) if last is not None else b
-                        c[i] = p
+                        if last is None:
+                            # For hue longer, we need to simulate the behavior of resolving undefined hue before fix-up
+                            if is_longer and hue_index == i:
+                                print('who?')
+                                c[i] = alg.lerp(b - mx, b, point * (e - 1))
+                            # Handle all other channels and hue approaches in the default way
+                            else:
+                                c[i] = b
+                        else:
+                            c[i] = alg.lerp(last, b, point * e)
 
                         # We just filled an alpha hole, premultiply the coordinates
                         if self.premultiplied and i == alpha:
@@ -205,9 +215,14 @@ class InterpolatorContinuous(Interpolator[AnyColor]):
             # Replace all undefined values that occurred prior to
             # finding the current defined value that have not been backfilled
             if backfill is not None and last is not None:
-                for c in coords[backfill:]:
-                    c[i] = last
-
+                point = (len(coords) - backfill)
+                for e, c in enumerate(coords[backfill:], 1):
+                    # For hue longer, we need to simulate the behavior of resolving undefined hue before fix-up
+                    if is_longer and hue_index == i:
+                        c[i] = alg.lerp(last, last + mx, point * e)
+                    # Handle all other channels and hue approaches in the default way
+                    else:
+                        c[i] = last
                     # We just filled an alpha hole, premultiply the coordinates
                     if self.premultiplied and i == alpha:
                         self.premultiply(c)